requirements.txt
Flask
gunicorn
twilio
pandas
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
openpyxl
